+++ src/main/java/com/example/giftrecommender/service/RecommendationSessionService.java
@@ -7,7 +7,7 @@ import com.example.giftrecommender.domain.entity.RecommendationSession;
 import com.example.giftrecommender.domain.enums.SessionStatus;
 import com.example.giftrecommender.domain.repository.GuestRepository;
 import com.example.giftrecommender.domain.repository.RecommendationSessionRepository;
-import com.example.giftrecommender.dto.RecommendationSessionRequestDto;
+import com.example.giftrecommender.dto.request.RecommendationSessionRequestDto;
 import com.example.giftrecommender.dto.response.RecommendationSessionResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
